[fix]: added & ran rubocop

+++ lib/stable/bootstrap.rb
@@ -1,4 +1,4 @@
-require "fileutils"
+require 'fileutils'
 
 module Stable
   module Bootstrap
@@ -7,25 +7,21 @@ module Stable
       FileUtils.mkdir_p(Paths.caddy_dir)
       FileUtils.mkdir_p(Paths.certs_dir)
 
-      unless File.exist?(Paths.apps_file)
-        File.write(Paths.apps_file, "--- []\n")
-      end
+      File.write(Paths.apps_file, "--- []\n") unless File.exist?(Paths.apps_file)
 
-      unless File.exist?(Paths.caddyfile)
-        File.write(Paths.caddyfile, "")
-      end
+      File.write(Paths.caddyfile, '') unless File.exist?(Paths.caddyfile)
 
       disable_rvm_autolibs!
     end
 
     def self.disable_rvm_autolibs!
-      return unless system("which rvm > /dev/null")
+      return unless system('which rvm > /dev/null')
 
       # Only run once
-      marker = File.join(Paths.root, ".rvm_autolibs_disabled")
+      marker = File.join(Paths.root, '.rvm_autolibs_disabled')
       return if File.exist?(marker)
 
-      puts "Configuring RVM (disabling autolibs)..."
+      puts 'Configuring RVM (disabling autolibs)...'
 
       system("bash -lc 'rvm autolibs disable'")
       system("bash -lc 'echo rvm_silence_path_mismatch_check_flag=1 >> ~/.rvmrc'")
